refactor(database): reorganize SSL-related classes and imports
