+++ lampray-web/src/main/java/tech/lamprism/lampray/system/database/DatabaseResourceCleanup.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright (C) 2023-2026 RollW
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package tech.lamprism.lampray.system.database
 
 internal fun closeDatabaseResources(resources: Iterable<AutoCloseable>): Exception? {
@@ -9,7 +25,7 @@ internal fun closeDatabaseResources(resources: Iterable<AutoCloseable>): Excepti
             if (failure == null) {
                 failure = e
             } else {
-                failure?.addSuppressed(e)
+                failure.addSuppressed(e)
             }
         }
     }
