+++ lampray-web/src/main/java/tech/lamprism/lampray/system/database/ssl/DatabaseSslArtifacts.kt
@@ -1,0 +1,40 @@
+/*
+ * Copyright (C) 2023-2026 RollW
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package tech.lamprism.lampray.system.database.ssl
+
+import java.nio.file.Path
+
+data class DatabaseSslArtifacts(
+    val properties: Map<String, String> = emptyMap(),
+    val resources: List<AutoCloseable> = emptyList()
+) {
+    companion object {
+        val EMPTY = DatabaseSslArtifacts()
+    }
+}
+
+data class DatabaseSslFileArtifact(
+    val path: Path,
+    val resources: List<AutoCloseable> = emptyList()
+)
+
+data class DatabaseSslKeyStoreArtifact(
+    val path: Path,
+    val password: String,
+    val type: String,
+    val resources: List<AutoCloseable> = emptyList()
+)
