+++ lampray-web/src/main/java/tech/lamprism/lampray/system/database/ssl/DatabaseSslConfig.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2023-2025 RollW
+ * Copyright (C) 2023-2026 RollW
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package tech.lamprism.lampray.system.database
+package tech.lamprism.lampray.system.database.ssl
 
 data class DatabaseSslConfig(
     val mode: DatabaseSslMode = DatabaseSslMode.DISABLED,
