+++ lampray-web/src/main/java/tech/lamprism/lampray/system/database/ssl/DatabaseSslParsing.kt
@@ -1,9 +1,27 @@
-package tech.lamprism.lampray.system.database
+/*
+ * Copyright (C) 2023-2026 RollW
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package tech.lamprism.lampray.system.database.ssl
 
 import org.bouncycastle.asn1.ASN1Sequence
 import org.bouncycastle.asn1.DERNull
-import org.bouncycastle.asn1.pkcs.PrivateKeyInfo
 import org.bouncycastle.asn1.pkcs.PKCSObjectIdentifiers
+import org.bouncycastle.asn1.pkcs.PrivateKeyInfo
+import org.bouncycastle.asn1.pkcs.RSAPrivateKey
+import org.bouncycastle.asn1.sec.ECPrivateKey
 import org.bouncycastle.asn1.x509.AlgorithmIdentifier
 import org.bouncycastle.asn1.x9.X9ObjectIdentifiers
 import org.bouncycastle.cert.X509CertificateHolder
@@ -134,7 +152,7 @@ private fun createPrivateKeyInfoFromLegacyDer(keyBytes: ByteArray): PrivateKeyIn
     val sequence = ASN1Sequence.getInstance(keyBytes)
 
     runCatching {
-        val rsaPrivateKey = org.bouncycastle.asn1.pkcs.RSAPrivateKey.getInstance(sequence)
+        val rsaPrivateKey = RSAPrivateKey.getInstance(sequence)
         PrivateKeyInfo(
             AlgorithmIdentifier(PKCSObjectIdentifiers.rsaEncryption, DERNull.INSTANCE),
             rsaPrivateKey
@@ -142,7 +160,7 @@ private fun createPrivateKeyInfoFromLegacyDer(keyBytes: ByteArray): PrivateKeyIn
     }.getOrNull()?.let { return it }
 
     runCatching {
-        val ecPrivateKey = org.bouncycastle.asn1.sec.ECPrivateKey.getInstance(sequence)
+        val ecPrivateKey = ECPrivateKey.getInstance(sequence)
         val parameters = ecPrivateKey.parametersObject
             ?: throw IllegalArgumentException("EC private key is missing named curve parameters.")
         PrivateKeyInfo(
